Size to 1D array: EXCEPT ObjectiveMechanism
This commit changes the "size" attribute to a 1D array for all processing mechanisms, and (in theory) allows all processing mechanisms to be initialized with "size" rather than with "default_input_value". Further tests will be added in a later commit.

This commit also adds size to the arguments of the constructors of ObjectiveMechanism and ComparatorMechanism, which previously did not have them.

Not sure why LCA has so many identical -/+ sections in the comments.