Test Updates: DDM and TransferMechanism
Adds tests of DDM and TransferMechanism involving initializing them with size as an argument.

NOTE: IntegratorMechanism and GatingMechanism tests have NOT been updated, because I wasn't sure how size would be used in initializing them. If size could be used in initializing them, we can add tests. If size definitely will not be used in initializing them, we could forbid it with an exception.