Fixed bug with reset_initializer
We wanted to add a line to set “reset_initializer” in Component’s default property setter: whenever a parameter state’s variable is changed using the property setter (i.e. when somebody calls myMech.matrix = m). This is a more general solution to the previous hack in MappingProjection.

when I added the “reset_initializer” line, it broke the code: when the AccumulatorIntegrator tried to execute, it found that the ‘previous_value’ attribute was a string: “AutoAssignMatrix”, so the multiplication failed.

When creating a Projection, the Projection (during _instantiate_attributes_before_function) creates parameter states based on its params and function_params

The _instantiate_parameter_states does this by PARSING the parameters, then placing them into the parameter state. In the case of matrix, it turns the “AutoAssignMatrix” keyword into the appropriate matrix.

Thus, the parameter state is initialized correctly, but the function_params have not been changed, so the function_params still say “matrix = ‘AutoAssignMatrix’”

Then, the init moves to _instantiate_function. When the Function sets its own parameters, it sets them based on function_params, which have not been parsed. Thus, the matrix gets overwritten to the string keyword ‘AutoAssignMatrix’, then when the Projection tries to execute its function, it gets an error.

Evidence: the function_params attribute of the Projection also shows ‘matrix’ as a keyword string

Why hasn’t this problem been discovered before?

Because adding the ‘reset_initializer’ causes the “setting” of the ‘matrix’ parameterState (which uses an AccumulatorIntegrator) to actually work. Previously, when the matrix parameter was being overwritten during _instantiate_function, nothing happened to it.

The matrix parameter is the only one that is parsed into a completely different format (from string to array) by _instantiate_parameter_states.

The solution is simply to tell the Projection's LinearMatrix Function to properly set the matrix (after parsing the matrix) and its parameter state and its parameter state's `previous_value` by calling the setter from make_property.