Kwta development (#391)
* Minor Changes

* Auto and Cross: v1.0

Added "auto" and "cross" arguments to RecurrentTransferMechnanism.

Due to the oddness of params, I had to add the "auto" and "cross" arguments to KWTA and LCA (child classes of RecurrentTransferMechnanism) as well.

* Testing Changes

Added a few tests to the tests for RecurrentTransferMechanism, and organized them a bit better.

* Minor Doc Change

* Removed print statements

* Added for AutoAssociativeProjection

Added the following:

AUTO_ASSOCIATIVE_PROJECTION = "AutoAssociativeProjection"
AUTO = 'auto'
CROSS = 'cross'

Removed the following:
AUTO = True

* Prototype AutoAssociativeProjection

* Typos and Bugfixes

* Minor Fixes

* Typo Corrections

* AutoAssociativeProjection v0.1

Primitive AutoAssociativeProjection

Lots of bugs. But shows the basic functionality. Havne't tested modulation but should in theory be working now.

* Minor Doc Changes

* Bugfix in show_graph

* Enable AutoAssociativeProjection to infer auto/cross from matrix

* Minor Typos

* Minor bugfixes

* Minor import fixes

* Move Auto and Cross to the Mechanism

Now Auto and Cross are owned by the RecurrentTransferMechanism rather than the AutoAssociativeProjection.

This makes it easier to modulate them.

* Minor Typos/Bugfixes

* Minor Changes

* Minor Changes

* Fixed bug with reset_initializer

We wanted to add a line to set “reset_initializer” in Component’s default property setter: whenever a parameter state’s variable is changed using the property setter (i.e. when somebody calls myMech.matrix = m). This is a more general solution to the previous hack in MappingProjection.

when I added the “reset_initializer” line, it broke the code: when the AccumulatorIntegrator tried to execute, it found that the ‘previous_value’ attribute was a string: “AutoAssignMatrix”, so the multiplication failed.

When creating a Projection, the Projection (during _instantiate_attributes_before_function) creates parameter states based on its params and function_params

The _instantiate_parameter_states does this by PARSING the parameters, then placing them into the parameter state. In the case of matrix, it turns the “AutoAssignMatrix” keyword into the appropriate matrix.

Thus, the parameter state is initialized correctly, but the function_params have not been changed, so the function_params still say “matrix = ‘AutoAssignMatrix’”

Then, the init moves to _instantiate_function. When the Function sets its own parameters, it sets them based on function_params, which have not been parsed. Thus, the matrix gets overwritten to the string keyword ‘AutoAssignMatrix’, then when the Projection tries to execute its function, it gets an error.

Evidence: the function_params attribute of the Projection also shows ‘matrix’ as a keyword string

Why hasn’t this problem been discovered before?

Because adding the ‘reset_initializer’ causes the “setting” of the ‘matrix’ parameterState (which uses an AccumulatorIntegrator) to actually work. Previously, when the matrix parameter was being overwritten during _instantiate_function, nothing happened to it.

The matrix parameter is the only one that is parsed into a completely different format (from string to array) by _instantiate_parameter_states.

The solution is simply to tell the Projection's LinearMatrix Function to properly set the matrix (after parsing the matrix) and its parameter state and its parameter state's `previous_value` by calling the setter from make_property.

* Fixed Initial_value of integrators, plus other Minor Fixes

* Update RecurrentTransferMechanism Tests

* Fixed Bug with RecurrentTransferMechanisms in dependency graphs

* Fixed Potential Bug with Learning

NOT YET TESTED THIS

Still need to add tests with learning and recurrent projections. The included code was just to fix a potential bug.

* Minor Fixes, More Tests

* added logger to KWTA

* Minor test update

* Fixed Bugs with setting auto/matrix

Moved _update_parameter_states to AFTER self.matrix is set in the execute() method of AutoAssociativeProjection.

Added quick-patch for fixing a bug with setting auto: the underlying problem of prev_context may not be solved.

* Minor Test Update and Bugfix in Process

Fixed an indexing typo in Process that caused self-recurrent mechanisms at the end of a pathway to create an error.

* KWTA Update (Medium-size)

Updated KWTA:

-expects the user to _never_ attempt to create stuff for the inhibition input state, and acts accordingly (appends a vector of zeros to variable, appends an input state, appends zero-vector to all input from execute(), etc)
-creates an AutoAssociativeProjection instead of a MappingProjection.
-calls _handle_size
-moved stuff to _validate_inhib
-fixed minor bugs

* Some KWTA tests

* Updated Recurrent Transfer Mechanism tests

added tests for process

* Minor Changes

* Changes

all instances of MappingProjection changed to AutoAssociativeProjection

matrix changed to be a property basd on auto and cross

* modernized LCA

fixed it up to use auto and cross

now it's practically a shell class because it's so simple

* cross --> hetero

changed all occurrences of 'cross' to 'hetero'

* removed last few mentions of cross

changed last mentions of cross to hetero

* Updated RecurrentTransferM and AutoAssociativeP

Medium size update:

-created and updated docstrings
-moved _instantiate_recurrent_projection into the RecurrentTransferMechanism class because that allows KWTA to overwrite it later
-Added validation of auto and hetero
-AutoAssocP updates the auto and hetero paramStates now: instead of RecurrentTransferM
-cleaned up some extraneous code
-added matrix, auto, hetero getters/setters to be more intuitively synced with their counterparts in RecurrentTransferM

* Tentative Core Dump on AutoAssocP and RecurTransferM

I did a core dump (plopped ideas onto page) of the ideas behind RecurrentTransferM and AutoAssociativeP sharing/syncing some parameters.

this core dump may not belong here, and perhaps should be moved)

* Minor Changes

* More KWTA tests

* Updated Tests to Reflect Improved Syncing

(title refers to the improved syncing between AutoAssocP and RecurTransferM mentioned in the medium size commit a moment ago)

* Basic Bugfixes, updates

basic bugfixing
added some param validation as well
and updated kwta to override _instantiate_recurrent_projection rather than _instantiate_attributes_after_function

* Minor Update