Component: _initialize_parameters: handle iter order of parameters with aliases
Skip over aliases and handle with main parameter.

If an alias was encountered before its source Parameter when iterating through
parameter names, and the value for the source was user-specified, this value
would be copied when it would have been directly assigned as the Parameter
value.

Visible in tests/functions/test_buffer.py::test_buffer_standalone_noise_function
The NormalDist function assigned as noise should not be copied because it has no
owner. When 'additive_param' was encountered first, the function would be
copied.