DDM.max_executions_before_finished to sys.maxsize
Default value of 1000 is arbitrary, should be unlimited in my
opinion. When time_step_size gets below 0.001 then 1000 steps
are often needed. Setting the sys.maxsize make its practically
unlimited for now. This is a hack though, maybe there is a way
to disable the limit but I can't find it in the docs.