github-actions: Prepare pre-release on tag on master
When a tag is pushed to master, this workloads creates a new
pre-release (which triggers the test-release workflow)

The default actions token does not trigger workflows on repo actions.
Creating a release thus needs PAT with public_repo privilege level
in order to trigger the test-release workflow.
If the token is set to 'no-token' the pre-release creation is skipped.

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>