functions/parameters: Reset PRNG state on seed assignment or change
Instead of resetting the PRNG state once per context,
the new approach compares the originally used seed with the new value
(either new setting or modulation).

This allows for better match in multi-trial executions with OCM:
Simulations will only reset PRNG if the modulated seed is different
from the one in the current execution context.
The result of evaluation will thus match the result of trial even
if OCM selects the same seed as originally used by the execution.

The PRNG state is also reset on direct seed modification to preserve
the old behaviour in situations like;
Func1.seed = n
Func2.seed = n
# Both Func1 and Func2 now follow the same random sequence irrespective
# of their original seed

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>