Update Cohen_Huston1994.py
Set back all the parameter values to the original settings as in the 1994 Cohen and Huston paper. Most importantly: now the weights before and after initialization are set correctly. I've also changed it so neutral input can be 1 (weight for neutral input was on 0), and I've changed it so that both word/color --> response and response --> word/color are changed before and after initialization. No idea what the rationale was from the original author how it was done, but for me it didn't make sense!