llvm, mechanism: Use base parameter structure to create all modulated parameters
The original approach created a copy to modulate mechanism parameters
(if needed to apply mech parameter ports) and then passed this copy to
internal function invocation.
Invoking internal functions would then create more copies (if needed)
to apply parameter ports of function parameters.

The new approach passes the mechanism base parameters,
so the copies of function parameters can be made from the original.

The overall amount of copied data is the same,
but now the same shared source is used for all copies

This is especially beneficial for GPUs that place the shared parameters
in high BW on-chip memories.

The observed effect for stability-flexibility model is ~20% reduction
in the total amount of data read from thread private memory,
Resulting in ~10% improvement in kernel execution time.
Measured on P620 GPU.

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>