Composition: add_projection: also detect feedback from node roles
Node roles may be specified in various places, including when specifying
a Pathway. These roles can imply that some projections should be
feedback, and it can be cumbersome to check for these in any place
add_projection may be called to pass feedback=True.

This checks for these roles in Composition.add_projection when adding an
edge to the Graph.

Fixes #2004