functions/SoftMax: Restore correct computation of derivation
Commit cae1465b319539f17bb651dd7961ede88a7c4022
	("llvm, functions/SoftMax: Implement compiled 'derivative' variant")
incorrectly assumed that the use of 'output' was an oversight.
It wasn't SoftMax derivative can take advantage of results if available.
This change restores the original functionality and adds a path to
compute the results if output is None.
This is used for testing where the results would need to be calculated
anyway.
The compiled variant is adapted in the same way, and the test are
updated to reflect the new results.

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>