Fix for getting correct subset of output ports
Use the port map on the inner composition and the outcome variables
parameter of PEC to determine the correct column indices of the simulation
results to pass to simulation_likelihood. Also added some better exceptions
and tests for checking when outcome_variables and data passed to PEC
constructor do not align.