Make DDM PEC test sensitive to incorrect inputs
The old test_parameter_estimation_ddm_mle passed had all inputs
set to 1. If the inputs were not getting passed to the composition
correctly the parameter estimation procedure would still recover
parameters because it wasn't impacting the results. In some cases
this was because it was using the last input value (stored from the
initial run of the inner composition during generating of the data to
fit). This single trial of input was being recycled by the compiled
code. In other cases, a single trial of value 0 was being used and
this also did not change the response either.

To test this more carefully, I have made the stimulus drift inputs to be
mostly negative so that most responses are negative. If the composition
does not receive the negative stimulus drift then it will generate a more
even response ratio. This causes the recovery of the DDM drift rate to be
skewed.

The current version of the inputs passing was broken so I reverted to the
hack of using the cached PEC inputs directly for now.