Fix bug caused by degenerate likelihood.
I had commented out a piece of code that checked
whether enough samples where aquired for the KDE
estimate. Since we are computing KDE over a joint
distribution of categorical (decision) and continuous
(reaction time) it can occur that a particular decision
value does not occur even once in all the simulations.
This makes the KDE estimation impossible and we must
impute a low but non-zero (zero will cause log to explode)
value for this probability. I had commented out this check
for some reason and it was leading to a crash.