Some changes to support PEC optimization in compiled mode
- Put back reservoir sampling approach in GridSearch._function. Jan
said this approach was used to maintain identical behaviour with the
end to end compiled version. To support PEC optimization in which
and aggregation function is specified we need to convert the ctypes
array to numpy to apply aggregation function. Now we then convert
this back to ctypes to run the LLVM min\max selection with reservoir
sampling.

- Added func_mode to test_pec args to test PEC optimization mode in
compiled mode.

- Marked test_pec in LLVM mode with objective_function=None as xfail.
Also improved the error message for that is returned a bit. Currently,
compiled mode doesn't seem to support PEC with no objective mechanism
unless we are in data fitting mode and returning the results of the
composition. This is probably easy to fix but I wanted to confirm with
Jan.

- Marked test_pec in LLVM mode with objective function set to
Concatenate as xfail. Concatenate doesn't seem to be supported in
compiled mode currently.