Composition: fix INPUT roles for ControlMechanism -> inner Composition
When a ControlMechanism controls something within an inner Composition,
a MappingProjection is created from the ControlMechanism to the inner
Composition's parameter CIM. This created a scheduling dependency in all
cases. This commit creates this dependency only when the
pathway/projection from the ControlMechanism to the inner Composition is
explicitly created by the user.