Fix test_processing_mechanism_function
The DriftDiffusionIntegrator was instantiating the Function
before passing to ProcessingMechamisn. This has the side-effect
of causing the Function within the ProcessingMechanism to return
a 1D shape rather than a 2D shape. For now, get rid of the
un-needed instantiation of the function. I think this is a bug
though so maybe the test should be re-written to try both cases
and make sure the output is always 2D from the functions when
placed inside a mechanism, regardless of how they are instantiated.