tests/AudodiffComposition: Add helper function to return the first set of learning results
The 'benchmark' fixture runs and arbitrary number of iterations.
Because of this it is override in conftest.py to preserve the return
value of the first invocation. However, this only applies to explicitly
returned values, not stateful variables.

Add a helper function that returns 'learning_results' to workaround
this restriction.

Fixes --benchmark-enable test in AutodiffComposition
Fixes: 521a4a2c773a8c12033352bbdbda62e3510d78b9 ("Tentative learning branch (#2623)")

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>