tests: test_call_after_construction_with_learning_pathway: fix missing assert
np.allclose call was not asserted since this test was added. The
results for this call changed in
0946af083f2f1cab23f0dfd6c0ee05b81b765500 (#1639) due to a change in
default for argument loss_spec (formerly loss_function) from None to
MSE in add_linear_learning_pathway and
_create_backpropagation_learning_pathway.

As of this commit, changing these defaults back to None produces the
numeric results that were present when this test was created.

The allclose call also misinterpreted the return value for
Composition.run as producing multiple trial outputs that are instead in
Composition.results