tests: ensure seed change test uses new seed
Since each test uses incrementing fixed seeds from 0, using new_seed = 1
will cause it to share a seed with some other component, depending on
the order in which components are created (ex: one Parameter is
instantiated before or after another, resulting in different individual
seeds for each).

This ensures that new_seed will actually be unused in the test.