Fix explicit returns mixed with implicit
Code scanning alerts caught this bug I had were the code
would have broke if display_iter was False. Should have
some coverage for this. It is always True for now but
down the line there might be an optimization algorithm
that doesn't want to display iteration progress for
some reason (wants to handle it on its own).