Fix bug in PEC mapping outcomes variables to data.
It turns out the that PEC code was assuming the inner
compositions port map dict order was always the same
as the output ports list. This was non-deterministic
it seems. Now we Get the index of the outcome variable
in the output ports of inner composition. To do this,
we must use the inner composition's portmap to get
the CIM output port that corresponds to the outcome
variable.