parameters: sync numeric values with compiled structures
On compilation (Execution._get_compilation_param), assign shared memory
(as numpy array) of compiled structures to values of Parameter objects
represented in llvm param and state structs. This allows the user to
make compatible changes to these values using Parameter.set without
recompilation.

Delete all compiled data for a given context when an incompatible change
is made to the value of any of these Parameter objects.