composition: Create a copy of 'controller_condition'
If the controller_condition is not set by the caller it will be
a global instance of the "Always" condition. Adding a reference
there results in a leak.
Creating a copy for use in controller_condition avoids adding
a reference to a global instance.

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>

Revert "composition: Do not the store owner of 'controller_condition'"

This reverts commit a9194e8176763648f5c3ffff55c202df756c75f0.