TransferMechanism, Component: Disable special setter logic when synchronizing compiled params
Synchronization with compiled params should only replace the current
value with a shared memory numpy array. The new numpy array should
also have the same value as the currently set one.
Do not propagate the new value.
Propagating the value could result in changes to other parameters or
overuse of shared memory arrays resulting in parameter value corruption.

Add explicit checks to test with custom 'has_initializers' setting.
The test currently passes in compiled mode only because scalar values
do not share memory with the compiled structure so the corrupted
parameters are not used in compiled execution.
Using shared memory for scalar values would lead to intermittent
failures depending on the order in which 'integrator_mode' and
'has_initializers' are synchronized.

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>