Fix PEC conditional test.
The PEC conditional test was still marked skip. Now it is fixed and passing.

I also lowered num_estimates on a lot of the PEC tests to make them faster. These are a low as I can go for the most part without having to change their expected test values.