Handle MappingProxyType in check_user_specified
If a user passes a MappingProxyType as argument
to a constructor, it gets cached in _user_specified_args. Later, if the object is
deepcopied, the MappingProxyType can't be pickled
so we get an error. I am handling this by converting the mapping proxy type to a normal dict.

This was uncovered by Jan's changes to tests that
always passes dictionary