pytorch: implement input port function execution
Previously, pytorch Mechansims would only use the equivalent of a
SUM LinearCombination for each of their input ports.

This implements the use in pytorch of the actual Functions assigned to
the input ports. When the pytorch-equivalent of a psyneulink Function
(via _gen_pytorch_fct) is unavailable, an error is thrown.