Add a real pytorch training test with batching
I made a test that implements an equivalent network
in PyTorch and PNL, initializes them with the same
weights, trains them both and then compares the losses. I think this is similar to some other
tests in the file but this one tests batch size > 1
and it actually trains the pytorch network as well.