Disable update_autodiff_all_output_values
Now that all_output_values contains the
batch dimension in the first dimension it
is causing and error when passed to
output_CIM.execute because it has the incorrect
number of input ports (unless batch size happens
to match number of input ports). Not sure what
the best fix to this is but disabling this line
seems to work and pass all tests. Need to ask Jon.