Add support for batching to AutoDiffCompostion in PyTorch mode. (#3204)
* Changes to allow passing BatchedImpl tensors

When pytorch vmap is used on a function, the function
is called to trace it with tenors passed as inputs
which are not normal pytorch tensors. They are in fact
BatchedImpl tensors that have no underlying storage.
These cannot be converted to double() or numpy arrays
so I needed to remove calls to convert these tensors
in this special case.

* Fix bug in _batch_inputs

_batch_inputs was not actually returning batches of
size batch_size.

* Testing bactch support.

* Fix batch_size=1 case.

* WIP merge for pytorch batching in autodiff

* Fix for when minibatch_size is ndarray

* Cleanup

* Remove handling of batched_results.

* Handle dtype-object in _get_variable_from_input

Not sure this works in general to be honest.

* Fixes from merging in devel.

* Fixes for ragged things.

* More fixes for ragged things.

* Fix bad torch.stack call in execute_input_ports

* Add a real pytorch training test with batching

I made a test that implements an equivalent network
in PyTorch and PNL, initializes them with the same
weights, trains them both and then compares the losses. I think this is similar to some other
tests in the file but this one tests batch size > 1
and it actually trains the pytorch network as well.

* More fixes

* Disable update_autodiff_all_output_values

Now that all_output_values contains the
batch dimension in the first dimension it
is causing and error when passed to
output_CIM.execute because it has the incorrect
number of input ports (unless batch size happens
to match number of input ports). Not sure what
the best fix to this is but disabling this line
seems to work and pass all tests. Need to ask Jon.

* Handle 3D structure better in autodiff update_results.

* Fix formatting

* Fix ragged processing.

A bunch of fixes for things wrong with ragged processing.

* Fix logging test.

We need to add batch dimension to the logging checks.

I also made the test check multiple batch sizes.

* Move variable list initialization outside look

I am pretty sure this was a bug.

* Fix ugly formatting

* Default variables need to be forced to atleast_2d

Since we have a batch dimension now, all the default
variable need to be forced to 2D.

* Fix store_memory to support batches.

Not sure this is the correct fix, I am getting errors
in results, but I don't know how else to implement
this.

* More fixes for defaults with batch size.

* More fixes for defaults with batch size.

* Fix store_memory

* Fix store_memory

* Uncomment some tests in test_emcomposition.py

* Revert change to _get_variable_from_input

I think I made this change in error, it was before
I fixed issues with handling ragged structures. Lets
see if tests pass.

* Execute the output_CIM on the last element of the batch to update the output ports

* Move check for torch to learn method.

Since we need torch now to setup inputs (before execute), I moved the check for it to learn.

* Add np.VisibleDeprecationWarning to catch.

Looks like old versions of numpy didn't throw ValueError

* Add pytorch mark to test

* Override new weights_only=True default in PyTorch 2.6

* Move the pytorch lower pin up to 1.13

Looks like weights_only argument wasn't
supported till then. That is a pretty ancient version of pytorch anyway.

Also pushed the upper pin up.

* Documentation cleanup

* Add a check for batch_size > 1 in EMSStorage

* CodeQL cleanup

* Better testing of batched_results argument.

* Style fixes.

---------

Co-authored-by: David Turner <dmturmer@princeton.edu>
Co-authored-by: jdcpni <jdc@princeton.edu>