Fix incorrect randomization seeding in PEC.
The initial_seed and same_seed_for_all_parameter_combinations arguments to the PEC were not being propagated correctly.

A new seed sequence was not being generated for each
run of simulations. In addition, the initial_seed of
the seed sequence was always 1. This fixes both of
these issues. However, I think we should probably
change how the seed sequence is generated as well.
Currently we just increment the start and stop of the seed sequence by 1 to get a new sequence. Maybe
it is better to use something like this:

https://numpy.org/doc/2.1/reference/random/parallel.html#seedsequence-spawn