Parameter: rework fallback for missing values in a Context
Previously, if a Parameter did not have a value for a given Context,
Parameter.get would return None or its default_value based on the value
of Parameter.fallback_default. Several places did not distinguish
between this type of None and a "no user-specified value, use a default"
None, which resulted in unexpected behavior or bugs.

In this commit, Parameter.fallback_default is replaced with
Parameter.fallback_value, which can be overridden for a single call by a
new argument to Parameter.get: fallback_value. If the Parameter has no
value for a Context, Parameter.get behaves according to fallback_value:
- ParameterNoValueError (default) - raises a ParameterNoValueError
- pnl.DEFAULT (keyword ('default')) - returns the Parameter's
default_value
- any other value - returns that value