control: fix nested Composition context bug (#3262)
Include Projection.sender in _dependent_components.

Fixes bug with control in nested Compositions, where the outer
Composition's ControlSignal to the inner Composition's Parameter CIM
would not be initialized for the inner Composition's
simulations.
This caused the afferents of such a signal to silently use a default
value instead of the controlled value.