ci/ga: Install SumatraPDF as default pdf viewer on windows runners
Windows runners use MS Edge as the default pdf viewer.
This causes issues if any of the tests calls show_graph()
or any other routine that by default produces a .pdf file
and opens is in the default application.
The problem with MS Edge is that it intermittently fails
to exit and hangs the entire CI job in the last step.
see: https://github.com/actions/runner/issues/3383
Removing the default .pdf file association on windows is
quite difficult because it needs to check multiple places
and elevated privileges.
Installing an alternative pdf viewer works around the issue
as it prevents MS Edge from running when .pdf is displayed.

Works around: https://github.com/actions/runner/issues/3383

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>