log: correct missing entries in dictionary with time_values
Running a logged Component outside of and between Composition.run calls
caused Log._assemble_entry_data to lose track of the Time objects
associated with the runs, because it assumed they would be subsequent,
or separated by at most one entry. This commit matches the requested
time_values with their associated log entries regardless of position in
the log.