Swap pycodestyle\pydocstyle for ruff
I had to disable a bunch of extra pydocstlye errors when using ruff.
A lot of these seem auto fixable though. There are also a ton
of ruff errors as well we can autofix. I will leave that till later though
since it will touch a massive amount of files.